search: Cleanup CodeMirror query input (part 1) (#57268)
Context/plan

We currently have to "main" query input implementations, the "old" one in client/branded/src/search-ui/input/CodeMirrorQueryInput.tsx and the experimental/new one in client/branded/src/search-ui/input/experimental/CodeMirrorQueryInputWrapper.tsx. They share quite a bit of code but not as much as they could.

There are also two "configurations" of the old query input, CodeMirrorMonacoFacade and CodeMirrorQueryInput.

All these differ from each other in two ways:

- which (CodeMirror) extensions configured
- how they integrate with other UI elements "around" them

My plan is to simplify at least the first one by rearranging the logic inside these components into reusable/configurable extensions that can be passed into a single "base" query input components. A lot of the functionality is already in extensions but more can be done here.
I think this will make things more manageable and easier to understand.

The purpose of the base query input is to provide the absolute minimal functionality for a query input, which is primarily syntax highlighting, default keybindings and other default styles.
More advanced features are then added as extension as needed, including but not limited to:

- active token decoration
- token hover information
- autocompletion
- query validation
- ...

By making these configurable we'll be able to provide a better UX in the context where the input is used (for example the query input for search contexts should only allow certain filters).


Changes in this PR

- Addition of the new BaseCodeMirrorQueryInput which provides syntax highlighting, styling and keyboard bindings.
- This new component is used by 
  - the existing CodeMirrorQueryInput component, which adds search query validation, token hover information and other styles on top
  - the experimental/new query input
- Removed styles from other components that are not provided in the base component
- Changed the activeFilterExtension to not highlight the active filter when the user is selecting text. This avoids conflicts between decoration and selection background colors.
- Removed &dark from all base themes and use the global .theme-dark selector instead. This lets us avoid the isLightTheme hook and theme changes are fully handled by CSS.
- Changed the implementation of the useCompartment hook (which was unused) to make it easier to use. It now returns the initial extension to use when initializing the editor and will handle updates by itself.

Test plan

- Inspected new stories for base query input
- Compared styles, text selection, token hover and autjocompletion behavior side by side for
  - main query input (old and new)
  - search context query input
  - notebooks query input