Gitserver: move subrepo permissions client to the gitserver client struct (#57429)
Most of our gitserver client methods require a subrepo permissions checker, but we require the caller to pass one in every time. This adds the checker to the gitserver client struct so that it can be used for all methods. Since it's one more thing that needs to be overridden for some tests, I added some more general helpers around constructing a test client as well.

Some history: the authz.DefaultSubrepoPermsClient was written as an global variable that was overridden during the enterprise init step. However, we no longer have an enterprise init step, so this does not need to be a global, mutable variable. This takes the first step towards making that not the case by removing many of the sites that reference the global authz.DefaultSubrepoPermsClient, consolidating all those references to the NewClient() method, which eventually should be made to require a subrepo perms client as a dependency rather than relying on the global.