search: remove search-hybrid flag (#57507)
In our last release (5.1) we actually made this feature flag unused, but
did not remove the final references to it. Additionally we add a
CHANGELOG entry for the removal to the 5.1 section. This is quite a
minor change, so I don't think its worth calling out even though 5.1 has
already been released.

Test Plan: CI