docs: don't mention "exhaustive" anymore
"Exhaustive" search and "Search Jobs" are overlapping features and it
can be confusing to have both mentioned in the docs.

"Exhaustive" was our first attempt to solve the customer need to run
searches over everything and get all results, but "Search Jobs" just
does it better.

We don't remove anything in the backend, because having the option of
"count: all" can still be useful, we just don't give it the name
"Exhaustive" anymore and we don't single it out as
dedicated feature.

Test plan:
CI