Search chore: tidy up dependencies on symbol search (#57576)
This contains a bunch of cleanups of search symbols code. Basically, the goal was to take all the global clients used by symbols and explicitly declare them in a new ZoektSymbolsClient.

In order to limit the scope of this change, I added a new DefaultZoektSymbolsClient so that the things that expect to be able to call this stuff without dependencies still can.