Perf: improve speed of `FetchCommits` (#57596)
The FetchCommits query is used to list the commit history of a file. That particular operation is quite slow, so we spent some time digging into why.

Traces indicated that there were a ton of followup requests to gitserver after resolving initial list of commits. We tracked this down to two sources:

- Do not do a followup request to resolve the root tree of a commit. We can just assume this exists.
- Do not resolve the parent commits for the sake of Perforce when we are not in a perforce repo.

Co-authored-by: Bolaji Olajide <25608335+BolajiOlajide@users.noreply.github.com>