Require approval to run Renovate
We don't seem to regularly address any of these PRs that are currently open, but Renovate force pushes to the PRs it currently has open. For some of the current branches, we ran what looks like >2500 builds so far.
That seems quite wasteful to me.

This makes the workflow explicit: You go to the dashboard, check the checkbox, and get the PR you want to merge in a moment.
No more crazy amount of CI jobs generated from pretty much stale PRs.

Also, most of the package specific rules felt outdated so removed them.

Once this is merged, I will close the current stale PRs (can be reopened using the dashboard) and we should ideally see less CI jobs.