Smart search: skip invalid generated queries (#57601)
In rare occasions, a smart search rule can generate an invalid query. Unlike
queries run through the search client, these are not validated. Downstream
search code can then panic as it parses and runs the query.

Now we validate the generated smart search query and move to the next rule if
validation fails. We could try to audit the smart search rules to ensure they
always produce valid queries, but that is complex and less robust.