codeintel: Fix path-matching query fragment for diagnostics in GraphQL (#57613)
Fixes #56904

We should use `LIKE` instead of `=` because the query fragment passed in
has the form `path + "%"` for [pattern matching](https://www.postgresql.org/docs/current/functions-matching.html#FUNCTIONS-LIKE).

That said, it seems a bit weird that the path used here is used as a pattern,
not as a literal path. When requesting other subfields like hovers,
the path is treated literally. 😕 Keeping that as-is for backwards compatibility,
and since nobody around has strong opinions on this.

## Test plan

Manually tested this against a fake index generated from scip-go modified
to emit diagnostics. I don't think it's a good idea to create more tech debt
by adding a large SQL file with a bunch of binary strings without clear
reproduction steps, which is what one of the other tests does for hovers.

(cherry picked from commit 4224ac23bacbb79eadc0578aa0e40342c14ce404)