search: use Zoekt parser in frontend (#57642)
search: keyword search for zoekt behind feature flag

We use Zoekt's query parser when we send a query to Zoekt, bypassing
Sourcegraph's current logic for interpreting patterns. Filters should
keep working as expected. The most notable difference is probably that
Zoekt interprets "foo bar" with an implicit "AND" while Sourcegraph
concatenates the two terms.

The intent is to experiment with different ways of parsing the user
query.

Test plan: CI

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>