goroutine: Don't silently stop after panic
If the handler function of a periodic goroutine panics, the handler silently dies. I don't see a panic error log, and it is never invoked again.

After discussing this on Slack, we concluded that periodic routines are most often part of a background job and that they shouldn't take down entire services when a single component misbehaves. Instead we treat them as an error in that component, which will surface in logs and the background jobs UI.

See https://sourcegraph.slack.com/archives/C02UC4WUX1Q/p1696960510011339 for the discussion to go with this behavior.