search: use 10x defaults if using Zoekt parser (#57692)
This change is behind a feature flag.

Our experiments showed that often we miss the top results especially if the keywords are very common. Increasing the limits for ShardMaxMatchCount and TotalMaxMatchCount improves the results. We use the same limits as we do for "keyword scoring", which are also identical to Zoekt's default limits.

I also renamed the internal feature flag from Keyword to UseZoektParser to avoid confusion with KeywordScoring.

Test plan:

local testing: ran a couple of searches with feat=search-new-keyword and confirmed results are looking promising