graphqlbackend/telemetry: fix privateMetadata type (#57719)
The `privateMetadata` field, which accepts arbitrary values, was using the wrong JSON type in the backend - it should be `JSONValue`, not `json.RawMessage`, which is another thing entirely. This causes the field to currently be unusable, as the backend type does not match with the GraphQL schema.

This bug does not affect any existing features - the new telemetry export is enabled, but no clients currently use the GraphQL mutation yet. However, it's important that this bug be fixed so that we can start using the new mutation in clients. In the meantime, we can do a clientside check to filter out this broken field (which is optional) until 5.2.2

(cherry picked from commit b55639d3cacf78cdaf7537ebe671330cbc341a13)