telemetrygateway: support arbitrary privateMetadata types from graphqlbackend (#57741)
The bug fix in #57719 (backported in #57720) fixed support for the common case, a JSON object `privateMetadata`, but the GraphQL schema describes that any valid JSON should be accepted, and GraphQL validation permits primitives like `string`:

![image](https://github.com/sourcegraph/sourcegraph/assets/23356519/657a023c-876a-46dc-a768-eba3af820ef2)

This change improves on #57719 with support for arbitrary types in case they get submitted. Similar to #57719, this is not a customer-affecting bug today as there are no usages, but it is important we land this fix to minimize the impact of this bug as we introduce usages like https://github.com/sourcegraph/cody/pull/1192.

The diff is mostly tests, as the fix is fairly minimal.

## Test plan

- [x] Updated unit tests assert that the GraphQL library we use does indeed accept incoming types as the equivalent Go type in `JSONValue.Value`
- [x] Updated unit tests assert that marshalling of arbitrary types to proto works as expected
- [x] Manual testing with `sg start` demonstrates GraphQL requests with arbitrary `privateMetadata` types are accepted:
![image](https://github.com/sourcegraph/sourcegraph/assets/23356519/946b9ca2-8afd-479a-833b-85478542454c)
![image](https://github.com/sourcegraph/sourcegraph/assets/23356519/59aea94a-02f5-4ad6-8d1e-633ee091dd37)
![image](https://github.com/sourcegraph/sourcegraph/assets/23356519/fe62e257-1c70-4e46-ab88-4eb445f9009e)

(cherry picked from commit a9f2ba5e463df6cca22bda455da75429ce2215ce)