Bump p4-fusion to version of our fork
We've forked the tool and are now ready to try some of the fixes out in a more production enviroment, so bundling this temporarily from a specific commit from the fork repo.

Next up, I'll bump it in the gitserver image and add the new `--noBaseCommit` flag to the perforce syncer.