fix passing multiple args to jest --testPathIgnorePatterns
To pass multiple ignore patterns to Jest, you need to use `--testPathIgnorePatterns pattern1 --testPathIgnorePatterns pattern2 ...`. Otherwise, `--testPathIgnorePatterns pattern1 pattern2` treats `pattern2` as a test pattern to run, not to ignore.

The consequence of this is that some npm scripts and CI steps were running tests that they didn't need to run.