build //client/web:bundle with NODE_ENV=production
Previously, the bundle was being built with NODE_ENV=development. This caused unnecessary code to exist in the production bundle (and caused a devtools console warning for users with the React devtools browser extension).

Now, the bundle is being built correctly with NODE_ENV=production. It also uses linked sourcemaps and simple minification (not obfuscation) for the prod bundle.