JetBrains: Cody: Encoding port in requestFrom param
Previously, the JetBrains plugin used two query parameters during onboarding

- `requestFrom=JETBRAINS` to inform the user that they are authenticating for a JetBrains IDE.
- `port=NUMBER`: to indicate which `http://localhost:NUMBER` URL to redirect to after authorization. Without the port number, we don't know how to send the access token to the local IDE.

This PR adds support to encode both parameters together as `requestFrom=JETBRAINS-PORT_NUMBER`. The motivation for this change is to solve https://github.com/sourcegraph/jetbrains/issues/19. We hit on an issue that GitHub SSO was not working when linking directly from the IDE, and we suspect that the root issue is related to the two parameter encoding. If this change does not work as intended, then we plan to revert the diff.