gomod: update zoekt and go-ctags
This has a bunch of ranking tweaks such as upranking markdown
files (hence the go-ctags bump). A side effect of this change is that
markdown headings will appear in symbol search results. I have a
suspicion this won't be that bad in practice and in fact may be
useful (eg the symbol sidebar when viewing a markdown file).

If this is a problem, we will likely revert that change in go-ctags and
then work on how to only have that information for search ranking.

https://github.com/sourcegraph/zoekt/compare/16e2ff8c98ee...0f21f325cc5d

- f39e6eb573 zoekt: add debug flag to show DebugScore output
- 081cd037fd boost graphql types in results
- 70f5dd3fcc score: always upscore symbol matches
- d8bfea1efb score: factors for headers in markdown
- 4f214152eb score: clean up debug output
- 328fcb7a50 nix: use go 1.21 and universal-ctags 6.0.0
- 7cc2872df4 nix: use tag for rev in ctags derivation
- d3fc0dcee8 score: remove repetition-boost
- c869a248a3 scoring: score methods and funcs the same
- ca7ee51ee7 scoring: show atom count in debug score
- 5bbf05d436 fix tests
- 0a17ccb25a scoring: reduce allocations for addScore
- d4ba942a02 gomod: update go-ctags and ctags in docker
- 0f21f325cc C-tags: use type def instead of type alias

Test Plan: CI