Clean up querying of repository topics (#57755)
Previously, querying repos by topic was pretty messy, especially after we added support for gitlab topics. We had to dig down into the metadata, which not particularly nicely structured JSON for github, and we had to do this both for the index and for queries.

This adds a generated column topics which isolates the complexity of extracting the topics from the metadata while maintaining normalization. It replaces the codehost-specific indexes with a single GIN index on that column so that querying for repos with a specific topic is both fast and easy.