perforce: Run fsck and repack after clone
The initial "clone" of a p4-fusion depot will only contain loose objects, since it doesn't care about repacking for simplicity.
This is good to keep p4-fusion non-complex, but bad for performance and disk usage, so instead we are now verifying the repos integrity (to make sure that if there's issues, we'll know they came from repack or p4-fusion) and then repack all the objects into a single pack.

### Test plan

Multiple manual test runs, checked repos remained valid, and that the progress report correctly renders.