env var DEV_WEB_BUILDER_NO_SPLITTING to disable code splitting
For local dev, you can set the env var `DEV_WEB_BUILDER_NO_SPLITTING` to disable code splitting in the esbuild web build. This produces a large single bundle, but it is ~25% faster to produce and means that page navigations need not fetch additional chunks (which incurs an esbuild full rebuild). The downside is that the bundle is quite large (~20MB), but parsing does not seem to be noticeably slow on my machine. So, this is an overall speedup for local dev.