sg: don't print output twice if build fails
This irritated me and I also think it's confusing to see the output
twice, once in red and once in default color.

The "easy" fix would be to change the underlying error and don't include
the output, but it turns out that `run.InRoot` is used in quite a few
things: `sg ci` helpers, migration helpers, etc. And I don't want to
suddenly stop output from appearing where it previously did.

So I special-case the case where we *do* have the output easily
available, which is where we check for the `installErr`.

It's a bit of a band-aid and ideally we'd fix the underlying issue, but
I also like errors having the output included by default because it's
helpful.