search: only use documentsRanks if jobs are enabled
With this change we won't use the document ranks for scoring if the jobs
to compute the ranks are not enabled.

We have noticed that the document ranks negatively impact the ranking in
our new search prototype on s2. My best guess is that the data is either
incomplete or outdated. Note that the jobs to compute the ranks are
currently disabled on all Cloud instances (incl. S2) an .com.

Test plan:
CI