gitserver: Add observability around ensureRevision
I would like to know how many calls we actually make, and with what outcome this function returns so added a prom metric for it.
This is for research purposes on how gitserver is used today, we can remove the metric later, or add it to a dashboard if it turns out to be super valuable for debugging.