search: only use documentsRanks if jobs are enabled (#57948)
With this change we won't use the document ranks for scoring if the jobs to compute the ranks are not enabled.

Why?

We have noticed that the document ranks negatively impact the ranking in our new search prototype. My best guess is that the data is either incomplete or outdated. Note, the jobs to compute the ranks are currently disabled on all Cloud instances (incl. S2) and .com.

Test plan:
updated unit test