search: override searchOptions w/ dev settings (#57892)
This lets us override zoekt.SearchOptions from Sourcegraph on a per-request basis for the current user. This only works for authenticated users on the stream endpoint IE interactive search. The purpose is to enable quick experimentation with different zoekt.SearchOptions (and maybe other settings in the future). The target audience is internal devs only.