Punish candidates with high ambiguity
Candidate ambiguity is a measure of how detailed the candidate SCIP in
comparison to ground truth.

When a candidate symbol has high ambiguity, it means that it occurs in a
lot of places where ground truth SCIP uses different symbols.

A demonstration of this method overloads in Java.
If you have 20 overloads of the same method (but with different
parameters), scip-java actually produces 20 different symbols (e.g.
"NodeRenderer#render(+19)").

Our current methods just produce a single symbol "NodeRenderer#render()"
for all those occurrences.

This commit penalises such occurrences by the logarithm of ambiguity.