References pane: fix typing on symbols return type (#58001)
When go marshals a nil slice to JSON, it marshals as null. However, unlike in go, null is not treated as an empty slice in javascript. This was causing type errors that showed up in the UI.

Rather than fixing this on the backend side of things (which would be fragile, since it's very conventional to use a nil slice to mean empty slice in go), this makes the field nullable in the typescript type.