search jobs: fix query highlighting
`SyntaxHighlightetSearchQuery` did not render regexp ranges correctly.
I couldn't find a simple fix so I decided to replace it with a read-only
CodeMirrorEditor.

I want to get this into the upcoming patch. However, we use
`SyntaxHighlightetSearchQuery` in a lot of places and it seems risky to
replace the component without baking it a bit on s2. Hence I only fix
this bug for Search Jobs.