blob: Refactor blame decorations
This commit refactors blame decorations to be more configuration driven
and to fix the following issues:

- Not working "lock first line" behavior
- Wrong (no) blame decoration shown after a folded range

The implementation is changed to only render a React node in a line
where we actually show a decoration. This should slighlty improve
performance. In order to still show the recency indicator across all
lines it is moved into its own gutter and rendered directly via the DOM.

Recency colors have been refactored to avoid having to use
`useIsLightTheme`.

`BlameColumn` wasn't used anymore and could be removed which also
allowed changes to the BlameDecoration` itself.