search jobs: fix query highlighting (#58022)
`SyntaxHighlightedSearchQuery` does not render regexp ranges, such as `[a-z]`, correctly. 
I couldn't find a simple fix so I decided to replace it with a read-only CodeMirror editor.

I want to get this into the upcoming patch. However, we use `SyntaxHighlightedSearchQuery` in a lot of places and it seems risky to replace the component without baking it a bit on s2. Hence I only fix this bug for Search Jobs.