gomod: update zoekt to 5.2 release branch
We create a new release branch in zoekt called "sourcegraph-5.2". It is
based off the version in the sourcegraph 5.2 release branch. We then
cherry-picked 3 commits from main into it. This is to avoid the large
number of changes we are making, but including the fixes we want.

https://github.com/sourcegraph/zoekt/compare/f17ff0bac96a...379ed5a0bb34

- 39e647cabf C-tags: use type def instead of type alias
- ebf3aedaee Ranking: standardize ctags kind names before scoring
- 379ed5a0bb Debug: make indexing timeout configurable

This update was done by running "./dev/zoekt/update sourcegraph-5.2"

Test Plan: CI already run in zoekt, then this CI.