Introduce normalised weighting of candidates
After computing the weights (using same jaccard measure) of individual
pairs of (candidate, ground truth) symbols, we collect all the ground
truth symbols that can be assigned to a given candidate, and normalise
the weights of each pair by dividing it by sum of all weights.

The idea behind this is to reassert the fact that mapping of symbols is
fuzzy, and therefore we shouldn't be selecting just 1 symbol - instead we spread
the fuzziness over all the occurrences, normalise them so they add up to
one.

Note that for a single alternative the weight will be 1, but that's not
a problem because even if some occurrences were missed, they will be
counted as part of false negatives, heavily discounting the effect of
this spurious 1.0 TP