code intel: Remove search context from search based code intel (#58010)
Using search context is problematic when viewing a file/repo that is not
part of the selected search context, which can happen when a custom
default context is set.

It doesn't seem necessary to include the context. The query I looked at
now doesn't have the context set anymore. Example:

    ^append$ type:symbol patternType:regexp count:50 case:yes file:\.(go)$ repo:^github.com/sourcegraph/sourcegraph$ index:only

I wasn't able to verify all code paths though.