do not show "repo last fetched" if it is recent (#58034)
This is the icon in the top right of each search result container that shows the last time the repo was fetched.

Now, it's only shown if the last-fetched date is 1 day or more in the past. If it is more recent than 1 day, the icon is not shown. This reduces visual noise and simplifies the UI.

![image](https://github.com/sourcegraph/sourcegraph/assets/1976/8237644d-491f-442d-a99b-2fafc7c53b9b)