Admin: add search indexing to repo menu (#58061)
When debugging a search issue, it's common to want to check the status of
search indexing. Currently, you need to open the repo menu and click through to
'Settings' -> 'Search Indexing', which is not obvious. This change adds 'Search
indexing' directly to the repo menu, matching what we do for 'Permissions' and
'Last sync log'.

To counterbalance this, this change also removes the 'Embeddings Jobs' item. It
felt cluttered to have two 'Embeddings' items, and we want to direct people
towards policies instead of manually scheduling jobs.