gitserver: more robust DiffSymbols (#58077)
We improve DiffSymbols to first check the commits exist. This is to
provide a better error message. Additionally we wrap the returned error
to include context of the command run.

We also improve hybrid search to fall back in the case a commit doesn't
exist. The most likely explanation is that there is some sort of race
and the indexed commit no longer exists in the repository. In that case
we should fall back to normal searcher.

Test Plan: CI. We do not explicitly test the error path here. However,
the happy path is covered by integration tests.